% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.POM.R
\name{summary.POM}
\alias{summary.POM}
\title{Summary method for POM objects}
\usage{
\method{summary}{POM}(object, ...)
}
\arguments{
\item{object}{An object of class \code{POM}.}

\item{...}{Additional arguments passed to other methods (unused).}
}
\value{
A list containing a single element: \code{dispersion}.
}
\description{
This S3 method provides a minimal summary object for custom null models...
}
