% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/null model.R
\name{summary.staar_nullmodel}
\alias{summary.staar_nullmodel}
\title{Summary method for staar_nullmodel}
\usage{
\method{summary}{staar_nullmodel}(object, ...)
}
\arguments{
\item{object}{An object of class \code{staar_nullmodel}.}

\item{...}{Additional arguments passed to summary.}
}
\value{
A summary list with dispersion parameter.
}
\description{
Summary method for staar_nullmodel
}
